Module reth_metrics::common::mpsc

Support for metering senders. Facilitates debugging by exposing metrics for number of messages sent, number of errors, etc.

Structs

MeteredPollSender

A wrapper type around PollSender that updates metrics on send.

MeteredReceiver

A wrapper type around Receiver that updates metrics on receive.

MeteredSender

A wrapper type around Sender that updates metrics on send.

UnboundedMeteredReceiver

A wrapper type around Receiver that updates metrics on receive.

UnboundedMeteredSender

A wrapper type around UnboundedSender that updates metrics on send.

Functions

metered_channel

Wrapper around mpsc::channel that returns a new bounded metered channel with the given buffer size.

metered_unbounded_channel

Wrapper around mpsc::unbounded_channel that returns a new unbounded metered channel.